+++ NEWS.md
@@ -1,5 +1,7 @@
 # geos (development version)
 
+* Dropped support for `wk::wksxp()`.
+
 # geos 0.0.1
 
 * Initial CRAN release.